refactor(server): unconditional composition-root teardown (#3502 PR3 step 2)
Extract three self-contained pieces out of server.ts verbatim:
- Static/SPA serving (buildPath, getAnalyticsScript, invalidateHtmlCache,
  the BASE_URL/root-deploy static+catch-all block) -> staticServing.ts,
  exposing configureStaticServing(app, deps?). invalidateHtmlCache is
  re-exported from server.ts as a safety net (grep confirms no current
  external importer).
- The global JSON.stringify BigInt patch -> utils/jsonBigIntReplacer.ts,
  imported for its module-load side effect.
- The auto-responder-trigger channel backfill migration ->
  services/autoResponderTriggerMigration.ts.

server.ts drops to 1,151 lines. Orphaned imports (fileURLToPath,
__filename/__dirname, robotsTxtHandler, createEmbedCspMiddleware,
generateAnalyticsScript/AnalyticsProvider, rewriteHtml) removed.
